Split field docs into its own trait `DocumentedFields`
